mobot/v0.3/requirements.txt
gradio
pyyaml
sentence-transformers
pymilvus
accelerate
dashscope
einops
transformers_stream_generator
auto-gptq
optimum
tiktoken